rename : 파일명 직관적으로 확인할 수 있도록 변경
